requirements.txt
dbfread==2.0.7
Flask==3.1.0
Flask-CORS==5.0.0
python-dotenv==1.0.1
APScheduler==3.11.0
Werkzeug==3.1.3
